LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/svg-script-href.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <svg><script href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=cc735297-1051-4304-af1f-964f61cb9cd1&amp;encodingcheck=&Gbreve;"></script></svg>


FAIL Speculative parsing, document.write(): svg-script-href Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
